Move IFontInfo into IGraphicsStructs.h
